Change hard-code Brad for hard-code Ben.
It also fixes a bug, where the script loop assumed that the I.dat from
the correlator was generated in the build folder. Where it depends from
where we are calling the script. We assumed here that we are calling it
from the source directory where script_SMC100_loop.py is.